+++ app/Models/Venue.php
@@ -86,4 +86,9 @@ class Venue extends Model
     {
         return $this->hasMany(Booking::class);
     }
+
+    public function users(): HasMany
+    {
+        return $this->hasMany(User::class);
+    }
 }